adds tokens folder
