+++ libraries/Pliant/ILexerRule.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Pliant.Tokens;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
