+++ libraries/Pliant/Nodes/TokenNode.cs
@@ -1,4 +1,6 @@
-﻿namespace Pliant.Nodes
+﻿using Pliant.Tokens;
+
+namespace Pliant.Nodes
 {
     public class TokenNode : ITokenNode
     {
