+++ libraries/Pliant/ParseEngine.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
+using Pliant.Tokens;
 
 namespace Pliant
 {
