+++ libraries/Pliant/RuleBuilder.cs
@@ -1,4 +1,5 @@
 ﻿using Pliant.Terminals;
+using Pliant.Tokens;
 using System;
 using System.Collections.Generic;
 using System.Linq;
