+++ libraries/Pliant/State.cs
@@ -1,9 +1,5 @@
 ﻿using Pliant.Nodes;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 
 namespace Pliant
 {
