+++ tests/Pliant.Tests.Unit/TerminalLexemeTests.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Pliant.Terminals;
+using Pliant.Tokens;
 using System;
 using System.Collections.Generic;
 using System.Linq;
